Jö schön, Threads u. Userprozesse werden ordnungsgemäß beendet Threads können schlafen, Mutex benutzt das jetzt auch Jetzt muß nur der Mutex auch überall verwendet werden

+++ userspace/start.c
@@ -1,11 +1,11 @@
 
-void syscall(int syscall_number, int arg0, int arg1, int arg2)
+void syscall(long syscall_number, long arg0, long arg1, long arg2)
 {
   	__asm__ __volatile__ (
 	   "int $0x80\n"
 	   :
 	   :"a" (syscall_number), "b" (arg0), "c" (arg1), "d" (arg2)
-  );  
+  );
 }
 
 void _start()
